@@ -1,3 +1,10 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2022-04-18 03:24:17
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2023-03-17 14:18:38
+ * @Description: 
+ */
 import 'package:flutter/material.dart';
 import 'package:flutter_demo_kit/flutter_demo_kit.dart';
 import './overlay_action_routes.dart';
@@ -5,9 +12,9 @@ import './overlay_action_routes.dart';
 import './overlay_action_util.dart';
 
 class TSOverlayActionHomePage extends CJTSBasePage {
-  final String title;
+  final String? title;
 
-  TSOverlayActionHomePage({Key key, this.title}) : super(key: key);
+  TSOverlayActionHomePage({Key? key, this.title}) : super(key: key);
 
   @override
 //  _CJTSTableHomeBasePageState createState() => _CJTSTableHomeBasePageState();
@@ -19,8 +26,6 @@ class TSOverlayActionHomePage extends CJTSBasePage {
 class _CQModulesHomePageState
     extends CJTSBasePageState<TSOverlayActionHomePage> {
   var sectionModels = [];
-
-  OverlayEntry _overlayEntry1;
 
   @override
   void dispose() {
