@@ -1,11 +1,18 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2022-04-18 03:24:17
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2023-03-17 14:19:29
+ * @Description: 
+ */
 import 'package:flutter/material.dart';
 import 'package:flutter_demo_kit/flutter_demo_kit.dart';
 import './overlay_util_routes.dart';
 
 class TSAlertUtilHomePage extends CJTSBasePage {
-  final String title;
+  final String? title;
 
-  TSAlertUtilHomePage({Key key, this.title}) : super(key: key);
+  TSAlertUtilHomePage({Key? key, this.title}) : super(key: key);
 
   @override
 //  _CJTSTableHomeBasePageState createState() => _CJTSTableHomeBasePageState();
