@@ -1,9 +1,16 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2022-04-18 03:24:17
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2023-03-17 15:13:04
+ * @Description: 
+ */
 import 'package:flutter/material.dart';
 import 'package:flutter_overlay_kit/alert/message_alert_view.dart';
 
 // ignore: camel_case_types
 class TSAlertView1_MessageAlertPage extends StatelessWidget {
-  const TSAlertView1_MessageAlertPage({Key key}) : super(key: key);
+  const TSAlertView1_MessageAlertPage({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -21,7 +28,7 @@ class TSAlertView1_MessageAlertPage extends StatelessWidget {
     );
   }
 
-  Widget _appBar() {
+  PreferredSizeWidget? _appBar() {
     return AppBar(
       title: Text('MessageAlertView 模块'),
     );
@@ -29,13 +36,13 @@ class TSAlertView1_MessageAlertPage extends StatelessWidget {
 
   Widget _pageWidget(context) {
     return Column(
-      // mainAxisAlignment: MainAxisAlignment.center,
+      mainAxisAlignment: MainAxisAlignment.start,
       crossAxisAlignment: CrossAxisAlignment.center,
       children: <Widget>[
         Text("信息弹窗"),
         SizedBox(height: 40),
         IKnowMessageAlertView(
-          title: "添加的图片数量超过限制",
+          title: "添加的图片数量3超过限制",
           message: "我是提示（我是一串很长的字符串我是一串很长的字符串我是一串很长的字符串我是一串很长的字符串）",
           iKnowTitle: "我知道了3",
           iKnowHandle: () {
