@@ -1,3 +1,10 @@
+/*
+ * @Author: dvlproad
+ * @Date: 2022-04-18 03:24:17
+ * @LastEditors: dvlproad
+ * @LastEditTime: 2023-03-17 14:15:46
+ * @Description: 
+ */
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_demo_kit/flutter_demo_kit.dart';
@@ -7,7 +14,13 @@ import './TSOverlayAction/overlay_action_home_page.dart';
 import './TSOverlayUtil/overlay_util_home_page.dart';
 
 class TSOverlayMainPage extends CJTSBaseTabBarPage {
-  TSOverlayMainPage({Key key}) : super(key: key);
+  TSOverlayMainPage({
+    Key? key,
+    List<dynamic>? tabbarModels,
+  }) : super(
+          key: key,
+          tabbarModels: tabbarModels,
+        );
 
   @override
 //  _TSOverlayMainPageState createState() => _TSOverlayMainPageState();
