@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_baseui_kit/flutter_baseui_kit.dart'; // 为了取button
+import 'package:flutter_baseui_kit/flutter_baseui_kit_adapt.dart';
 
 /*
  * 有间隔的'取消'+'确定' Buttons
@@ -8,6 +9,7 @@ class AlertdCancelOKSpacedButtons extends StatelessWidget {
   final double? height;
 
   final String cancelTitle;
+  final ThemeStateBGType cancelStyleType;
   final void Function() cancelHandle;
   // final Void Function(Void) cancelHandle;
 
@@ -18,6 +20,7 @@ class AlertdCancelOKSpacedButtons extends StatelessWidget {
     Key? key,
     this.height,
     this.cancelTitle = '取消',
+    this.cancelStyleType = ThemeStateBGType.theme_gray,
     required this.cancelHandle,
     this.okTitle = '确认',
     required this.okHandle,
@@ -37,29 +40,30 @@ class AlertdCancelOKSpacedButtons extends StatelessWidget {
   }
 
   Widget _row() {
-    double buttonHeight = this.height ?? 40;
+    double buttonHeight = this.height ?? 36.w_pt_cj;
 
     return Row(
-      mainAxisAlignment: MainAxisAlignment.spaceAround,
+      mainAxisAlignment: MainAxisAlignment.center,
       crossAxisAlignment: CrossAxisAlignment.center,
       children: <Widget>[
         ThemeBorderButton(
-          width: 100,
+          width: 116.w_pt_cj,
           height: buttonHeight,
-          borderColorType: ThemeBGType.theme,
+          borderColorType: this.cancelStyleType,
           title: this.cancelTitle,
-          titleStyle: ButtonBoldTextStyle(fontSize: 15.0),
-          cornerRadius: 20,
+          titleStyle: ButtonBoldTextStyle(fontSize: 14.w_pt_cj),
+          cornerRadius: 18.w_pt_cj,
           enable: true, // 不设置,默认true
           onPressed: this.cancelHandle,
         ),
+        Container(width: 10.w_pt_cj),
         ThemeBGButton(
-          width: 100,
+          width: 116.w_pt_cj,
           height: buttonHeight,
           bgColorType: ThemeBGType.theme,
           title: this.okTitle,
-          titleStyle: ButtonBoldTextStyle(fontSize: 15.0),
-          cornerRadius: 20,
+          titleStyle: ButtonBoldTextStyle(fontSize: 14.w_pt_cj),
+          cornerRadius: 18.w_pt_cj,
           enable: true, // 不设置,默认true
           onPressed: () {
             Future.delayed(Duration(milliseconds: 100), () {
@@ -92,8 +96,8 @@ class AlertIKnowSpacedButton extends StatelessWidget {
     return Container(
       child: Column(
         children: <Widget>[
-          Divider(color: Colors.grey, height: 1),
-          Container(height: 50, child: _row())
+          // Divider(color: Colors.grey, height: 1),
+          Container(height: 40, child: _row())
         ],
       ),
     );
@@ -103,9 +107,12 @@ class AlertIKnowSpacedButton extends StatelessWidget {
     return Row(
       mainAxisSize: MainAxisSize.min,
       children: <Widget>[
-        FlatButton(
-          // color: color,
-          // shape: StadiumBorder(),
+        ThemeBGButton(
+          title: this.iKnowTitle,
+          bgColorType: ThemeBGType.theme,
+          height: 40,
+          width: 240,
+          cornerRadius: 20,
           onPressed: () {
             Future.delayed(Duration(milliseconds: 100), () {
               if (this.iKnowTitle != null) {
@@ -113,12 +120,23 @@ class AlertIKnowSpacedButton extends StatelessWidget {
               }
             });
           },
-          child: Text(
-            this.iKnowTitle,
-            textAlign: TextAlign.center,
-            style: TextStyle(fontSize: 16.0, color: Colors.red),
-          ),
-        )
+        ),
+        // FlatButton(
+        //   // color: color,
+        //   // shape: StadiumBorder(),
+        //   onPressed: () {
+        //     Future.delayed(Duration(milliseconds: 100), () {
+        //       if (this.iKnowTitle != null) {
+        //         this.iKnowHandle();
+        //       }
+        //     });
+        //   },
+        //   child: Text(
+        //     this.iKnowTitle,
+        //     textAlign: TextAlign.center,
+        //     style: TextStyle(fontSize: 16.0, color: Colors.white),
+        //   ),
+        // )
       ],
     );
   }
