@@ -3,19 +3,26 @@ import './components/alert_closed_buttons.dart';
 import './components/alert_spaced_buttons.dart';
 import './components/flex_width_buttons.dart';
 import './components/alert_container.dart';
+import 'package:flutter_baseui_kit/flutter_baseui_kit.dart';
+import 'package:flutter_baseui_kit/flutter_baseui_kit_adapt.dart';
 
 class CJBaseMessageAlertView extends StatelessWidget {
+  final double? height;
+
   final String? title;
   final String? message;
   final TextAlign? messageAlign;
   final bool isCloseButton; // button按钮是否是紧密的，如果是的话底部间隔是0
+  final Function? closeHandle;
 
   CJBaseMessageAlertView({
     Key? key,
+    this.height,
     this.title,
     this.message,
     this.messageAlign,
     this.isCloseButton = false,
+    this.closeHandle,
   });
 
   Widget renderButtons() {
@@ -28,7 +35,7 @@ class CJBaseMessageAlertView extends StatelessWidget {
     Map allAlertMarginVertical = {
       // alert 竖直上的间距
       "title_buttons": [30.0, 30.0, 20.0, buttonToBottomDistance],
-      "title_message_buttons": [20.0, 10.0, 20.0, buttonToBottomDistance],
+      "title_message_buttons": [44.0, 10.0, 42.0, buttonToBottomDistance],
       "message_buttons": [20.0, 20.0, 0.0, buttonToBottomDistance],
     };
 
@@ -66,8 +73,10 @@ class CJBaseMessageAlertView extends StatelessWidget {
           this.title ?? '',
           textAlign: TextAlign.center,
           style: TextStyle(
-            fontSize: 18,
-            color: Colors.black,
+            fontFamily: 'PingFang SC',
+            fontWeight: FontWeight.bold,
+            fontSize: 16.f_pt_cj,
+            color: Color(0xff333333),
             decoration: TextDecoration.none,
           ),
         ),
@@ -84,8 +93,8 @@ class CJBaseMessageAlertView extends StatelessWidget {
           this.message ?? '',
           textAlign: messageAlign,
           style: TextStyle(
-            fontSize: 16,
-            color: Colors.grey[600],
+            fontSize: 13.f_pt_cj,
+            color: Color(0xff8b8b8b),
             decoration: TextDecoration.none,
           ),
         ),
@@ -105,17 +114,15 @@ class CJBaseMessageAlertView extends StatelessWidget {
 
     double buttonHeight = 40;
     double buttonsWidgetHeight = marginTop + buttonHeight + marginBottom;
-
     return CQAlertContainer(
-      contentWidget: Column(
-        children: columnWidgets,
-      ),
+      height: height,
+      contentWidget: Column(children: columnWidgets),
       buttonsWidgetHeight: buttonsWidgetHeight,
       buttonsWidget: Container(
-        // color: Colors.pink,
         padding: EdgeInsets.only(top: marginTop, bottom: marginBottom),
         child: renderButtons(),
       ),
+      closeHandle: closeHandle,
     );
   }
 }
@@ -129,6 +136,7 @@ class IKnowMessageAlertView extends CJBaseMessageAlertView {
 
   IKnowMessageAlertView({
     Key? key,
+    double? height,
     String? title,
     String? message,
     TextAlign? messageAlign,
@@ -136,6 +144,7 @@ class IKnowMessageAlertView extends CJBaseMessageAlertView {
     required this.iKnowHandle,
   }) : super(
           key: key,
+          height: height,
           title: title,
           message: message,
           messageAlign: messageAlign,
@@ -143,7 +152,7 @@ class IKnowMessageAlertView extends CJBaseMessageAlertView {
 
   @override
   Widget renderButtons() {
-    return AlertIKnowCloseButton(
+    return AlertIKnowSpacedButton(
       iKnowTitle: this.iKnowTitle,
       iKnowHandle: () {
         // print("点击Alert按钮:'我知道了'");
@@ -160,23 +169,29 @@ class IKnowMessageAlertView extends CJBaseMessageAlertView {
  */
 class CancelOKMessageAlertView extends CJBaseMessageAlertView {
   final String cancelTitle;
+  final ThemeStateBGType cancelStyleType;
   final Function? cancelHandle;
 
   final String okTitle;
   final Function? okHandle;
+  final Function? closeHandle;
 
   CancelOKMessageAlertView({
     Key? key,
+    double? heiht,
     String? title,
     String? message,
     TextAlign? messageAlign,
     this.cancelTitle = '取消',
+    this.cancelStyleType = ThemeStateBGType.theme_gray,
     this.cancelHandle,
+    this.closeHandle,
     this.okTitle = '确认',
     this.okHandle,
   })  : assert(title != null || message != null), // 不同同时为空
         super(
           key: key,
+          height: heiht,
           title: title,
           message: message,
           messageAlign: messageAlign,
@@ -186,6 +201,7 @@ class CancelOKMessageAlertView extends CJBaseMessageAlertView {
   Widget renderButtons() {
     return AlertdCancelOKSpacedButtons(
       cancelTitle: this.cancelTitle,
+      cancelStyleType: this.cancelStyleType,
       cancelHandle: () {
         // print("点击Alert按钮:'取消'");
         if (this.cancelHandle != null) {
